Anneal Reaction
Created an annealing reaction method for primer binding. More
sophisticated primer binding and dna annealing may be added later (such
as mismatching, dna/rna annealing, crispr/gRNA annealing, etc.)
